+++ src/rhodes/structures.py
@@ -1,33 +1,35 @@
 """Other structures for Rhodes."""
-import attr
+from typing import Union
 
-from .choice_rules import Not, derive_rule
+import attr
+import jsonpath_rw
+from attr.validators import instance_of
+
+__all__ = ("JsonPath", "convert_to_json_path")
+
+
+def _convert_path(value: Union[str, jsonpath_rw.JSONPath, "JsonPath"]) -> jsonpath_rw.JSONPath:
+    if isinstance(value, jsonpath_rw.JSONPath):
+        return value
+
+    if isinstance(value, JsonPath):
+        return value.path
+
+    return jsonpath_rw.parse(value)
 
 
 @attr.s(eq=False, order=False)
-class Variable:
+class JsonPath:
     """Represents a JSONPath variable in request/response body."""
 
-    # TODO: Validate the path is a valid JSONPath
-    path = attr.ib()
+    path: jsonpath_rw.JSONPath = attr.ib(validator=instance_of(jsonpath_rw.JSONPath), converter=_convert_path)
 
-    # TODO: Add __and__ and __or__ behaviors?
+    def __str__(self):
+        return str(self.path)
 
-    def __lt__(self, other):
-        return derive_rule(variable=self.path, operator="<", value=other)
 
-    def __le__(self, other):
-        return derive_rule(variable=self.path, operator="<=", value=other)
+def convert_to_json_path(value) -> JsonPath:
+    if isinstance(value, JsonPath):
+        return value
 
-    def __eq__(self, other):
-        return derive_rule(variable=self.path, operator="==", value=other)
-
-    def __ne__(self, other):
-        inner_rule = derive_rule(variable=self.path, operator="==", value=other)
-        return Not(Rule=inner_rule)
-
-    def __gt__(self, other):
-        return derive_rule(variable=self.path, operator=">", value=other)
-
-    def __ge__(self, other):
-        return derive_rule(variable=self.path, operator=">=", value=other)
+    return JsonPath(value)