+++ test/unit/test_converters.py
@@ -1,0 +1,33 @@
+"""Unit test suite for ``rhodes._converters``."""
+from enum import Enum
+
+import pytest
+import jsonpath_rw
+
+from rhodes._converters import convert_to_json_path
+from rhodes.structures import JsonPath
+from rhodes.choice_rules import VariablePath
+
+pytestmark = [pytest.mark.local, pytest.mark.functional]
+
+
+RAW_PATH = "$.path.to.value"
+
+
+class MyEnum(Enum):
+    MY_PATH = RAW_PATH
+
+
+@pytest.mark.parametrize("path_value", (
+    pytest.param(RAW_PATH, id="string"),
+    pytest.param(MyEnum.MY_PATH, id="enum"),
+    pytest.param(jsonpath_rw.parse(RAW_PATH), id="jsonpath_rw.JSONPath"),
+    pytest.param(JsonPath(RAW_PATH), id="JsonPath"),
+    pytest.param(VariablePath(RAW_PATH), id="VariablePath"),
+))
+def test_convert_to_json_path(path_value):
+    expected = JsonPath(RAW_PATH)
+
+    test = convert_to_json_path(path_value)
+
+    assert test.path == expected.path