+++ CHANGELOG.rst
@@ -5,8 +5,12 @@ Changelog
 0.2.0 -- 2019-11-??
 ===================
 
-* Rename ``ChoiceRule.then_`` to ``ChoiceRule.then``
+* **BREAKING CHANGE:** Rename ``ChoiceRule.then_`` to ``ChoiceRule.then``
   `#12 <https://github.com/mattsb42/rhodes/issues/12>`_
+* **BREAKING CHANGE:** Rework ``Variable`` into ``JsonPath`` and ``VariablePath``
+  `#3 <https://github.com/mattsb42/rhodes/issues/3>`_
+  `#10 <https://github.com/mattsb42/rhodes/issues/10>`_
+  `#11 <https://github.com/mattsb42/rhodes/issues/11>`_
 
 0.1.0 -- 2019-11-14
 ===================